spyder-devsphere/requirements.txt
spyder>=4.0.0b3pytest
pytest-qt